+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Add support for PHP v7.1 (temporary measure to move away from PHP v5).
 - Type declarations have been added to all parameters and return types.
+- Validation with `\InvalidArgumentException` for the `base_convert()`
+  parameters to match the warnings added for the original PHP function in v7.4.
+  - `$number` must be a string using only *alphanumeric* characters.
+  - `$fromBase` and `$toBase` must be an integer between *2* and *36*.
 ### Removed
 - **BC break**: Removed support for PHP versions < v7.1 as they are no longer
   [actively supported](https://php.net/supported-versions.php) by the PHP project.